moved Footer component into a separate folder
