+++ src/components/index.js
@@ -6,5 +6,5 @@ export { default as About } from "./About";
 export { default as Portfolio } from "./Portfolio";
 export { default as Home } from "./Home";
 export { default as Resume } from "./Resume/Resume";
-export { default as Footer } from "./Footer";
+export { default as Footer } from "./Footer/Footer";
 export { default as Blogs } from "./Blogs";
