+++ src/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
 import { HashRouter as Router } from "react-router-dom";
 import App from "./App";
